Fixed issues raised in pull request comments
