+++ src/content/app/species-selector/views/species-selector-search-results-view/SpeciesSelectorSearchResultsView.test.tsx
@@ -23,6 +23,7 @@ import RouteChecker from 'tests/router/RouteChecker';
 import type { CommittedItem } from 'src/content/app/species-selector/types/committedItem';
 import type { SearchResults } from 'src/shared/types/search-api/search-results';
 import { Strand } from 'src/shared/types/core-api/strand';
+import { createSelectedSpecies } from 'tests/fixtures/selected-species';
 
 const mockUseAppSelector = vi.fn();
 const useLazySearchGenesQuery = vi.fn();
@@ -40,14 +41,12 @@ vi.mock('src/shared/state/api-slices/searchApiSlice', () => ({
 }));
 
 const committedSpecies: CommittedItem[] = [
-  {
+  createSelectedSpecies({
     genome_id: 'homo_sapiens_GCA_000001405_29',
     genome_tag: 'human',
     common_name: 'Human',
     scientific_name: 'Homo sapiens',
     species_taxonomy_id: '9606',
-    type: null,
-    is_reference: true,
     assembly: {
       accession_id: 'GCA_000001405.29',
       name: 'GRCh38'
@@ -55,10 +54,8 @@ const committedSpecies: CommittedItem[] = [
     release: {
       name: 'Sep 2025',
       type: 'integrated'
-    },
-    isEnabled: true,
-    addedAt: 1
-  }
+    }
+  })
 ];
 
 const geneResults: SearchResults = {
