+++ src/content/search/views/search-results-view/SearchResultsView.test.tsx
@@ -25,6 +25,7 @@ import {
   getCommittedSpecies,
   getHasLoadedStoredSpecies
 } from 'src/content/app/species-selector/state/species-selector-general-slice/speciesSelectorGeneralSelectors';
+import { createSelectedSpecies } from 'tests/fixtures/selected-species';
 
 const mockUseAppSelector = vi.fn();
 const mockNavigate = vi.fn();
@@ -63,14 +64,12 @@ vi.mock('src/shared/state/api-slices/searchApiSlice', () => ({
 }));
 
 const committedSpecies: CommittedItem[] = [
-  {
+  createSelectedSpecies({
     genome_id: 'homo_sapiens_GCA_000001405_29',
     genome_tag: 'human',
     common_name: 'Human',
     scientific_name: 'Homo sapiens',
     species_taxonomy_id: '9606',
-    type: null,
-    is_reference: true,
     assembly: {
       accession_id: 'GCA_000001405.29',
       name: 'GRCh38'
@@ -78,10 +77,8 @@ const committedSpecies: CommittedItem[] = [
     release: {
       name: 'Sep 2025',
       type: 'integrated'
-    },
-    isEnabled: true,
-    addedAt: 1
-  }
+    }
+  })
 ];
 
 const geneResults: SearchResults = {
