+++ src/shared/components/app-icon/FeatureSearchIcon.tsx
@@ -16,7 +16,7 @@
 
 import classNames from 'classnames';
 
-import SVGIcon from 'static/icons/icon_feature_search.svg';
+import SVGIcon from 'static/icons/icon_search.svg';
 
 import styles from './AppIcon.module.css';
